chore: all slack comms moved to commerce-server and recordNewPurchase
