+++ apps/testingaccessibility/src/lib/modules.ts
@@ -1,4 +1,4 @@
-import {getPurchasedProduct} from 'server/get-purchased-product'
+import {getPurchasedProduct} from 'lib/get-purchased-product'
 import {sanityClient} from 'utils/sanity-client'
 import {SanityDocument} from '@sanity/client'
 import {NextApiRequest} from 'next'
