+++ apps/testingaccessibility/src/pages/api/stripe/webhook.ts
@@ -1,15 +1,13 @@
 import {buffer} from 'micro'
 import type {NextApiRequest, NextApiResponse} from 'next'
-import {sendServerEmail} from '@skillrecordings/skill-api'
+import {postSaleToSlack, sendServerEmail} from '@skillrecordings/skill-api'
 import {nextAuthOptions} from '../auth/[...nextauth]'
-import {recordNewPurchase} from '../../../utils/record-new-purchase'
 import {withSentry} from '@sentry/nextjs'
 import * as Sentry from '@sentry/nextjs'
 import {tracer, setupHttpTracing} from '@skillrecordings/honeycomb-tracer'
-import {stripe} from '@skillrecordings/commerce-server'
+import {stripe, recordNewPurchase} from '@skillrecordings/commerce-server'
 import {tagPurchaseConvertkit} from '../../../server/tag-purchase-convertkit'
 import {updatePurchaseStatusForCharge} from '../../../lib/purchases'
-import {postSaleToSlack} from '../../../server/post-to-slack'
 import {PurchaseStatus} from '@skillrecordings/skill-api'
 import {prisma} from '@skillrecordings/database'
