+++ apps/testingaccessibility/src/pages/team/index.tsx
@@ -3,7 +3,7 @@ import Layout from 'components/app/layout'
 import {find, get, isNull, isString} from 'lodash'
 import {convertToSerializeForNextResponse} from '@skillrecordings/commerce-server'
 import {GetServerSideProps} from 'next'
-import {getPurchasedProduct} from 'server/get-purchased-product'
+import {getPurchasedProduct} from 'lib/get-purchased-product'
 import InviteTeam from 'components/team'
 import {UserGroupIcon} from '@heroicons/react/outline'
 import Link from 'next/link'
