+++ packages/commerce-server/src/index.ts
@@ -7,3 +7,4 @@ export * from './validate-coupon'
 export * from './get-coupon-for-code'
 export * from './prisma-next-serializer'
 export * from './props-for-commerce'
+export * from './record-new-purchase'
