+++ packages/skill-api/package.json
@@ -47,7 +47,8 @@
     "preact": "^10.7.2",
     "preact-render-to-string": "^5.2.0",
     "sanitize-html": "^2.7.0",
-    "uuid": "^8.3.2"
+    "uuid": "^8.3.2",
+    "pluralize": "^8.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.16.0",
@@ -67,6 +68,7 @@
     "@types/node": "^17.0.0",
     "@types/nodemailer": "^6.4.4",
     "@types/oauth": "^0.9.1",
+    "@types/pluralize": "^0.0.29",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@types/sanitize-html": "^2.6.2",
