+++ packages/skill-api/process.d.ts
@@ -3,5 +3,6 @@ declare namespace NodeJS {
     STRIPE_SECRET_TOKEN: string
     EMAIL_SERVER_PORT: number
     NEXT_PUBLIC_DEFAULT_PRODUCT_ID: string
+    SLACK_ANNOUNCE_CHANNEL_ID: string
   }
 }
